+++ src/main/java/com/imsweb/mph/MphInput.java
@@ -64,6 +64,10 @@ public class MphInput {
         return year < 2001 && GroupUtility.validateBehavior(_behaviorIcdO2) ? _behaviorIcdO2 : _behaviorIcdO3;
     }
 
+    public String getIcdCode() {
+        return getHistology() + "/" + getBehavior();
+    }
+
     public String getLaterality() {
         return _laterality;
     }