+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Fixed a bug in 2018 head and neck rules.
 
+- Implemented 2021 updates of solid tumor rules.
+
 **Changes in version 1.21**
 
 - Updated hematopoietic and lymphoid data.